+++ libraries/session/src/androidTest/java/androidx/media3/session/SimpleBitmapLoaderTest.java
@@ -26,14 +26,19 @@ import androidx.test.core.app.ApplicationProvider;
 import androidx.test.ext.junit.runners.AndroidJUnit4;
 import com.google.common.util.concurrent.ListenableFuture;
 import com.google.common.util.concurrent.MoreExecutors;
+import java.io.File;
 import java.io.IOException;
 import java.net.MalformedURLException;
+import java.nio.file.Files;
+import java.nio.file.Paths;
 import java.util.concurrent.ExecutionException;
 import okhttp3.mockwebserver.MockResponse;
 import okhttp3.mockwebserver.MockWebServer;
 import okio.Buffer;
+import org.junit.Rule;
 import org.junit.Test;
 import org.junit.function.ThrowingRunnable;
+import org.junit.rules.TemporaryFolder;
 import org.junit.runner.RunWith;
 
 /**
@@ -46,6 +51,8 @@ import org.junit.runner.RunWith;
 public class SimpleBitmapLoaderTest {
 
   private static final String TEST_IMAGE_PATH = "media/jpeg/non-motion-photo-shortened.jpg";
+
+  @Rule public final TemporaryFolder tempFolder = new TemporaryFolder();
 
   @Test
   public void loadData() throws Exception {
@@ -70,11 +77,13 @@ public class SimpleBitmapLoaderTest {
     ListenableFuture<Bitmap> future = bitmapLoader.decodeBitmap(new byte[0]);
 
     assertException(
-        future::get, IllegalArgumentException.class, /* messagePart= */ "Could not decode bitmap");
+        future::get,
+        IllegalArgumentException.class,
+        /* messagePart= */ "Could not decode image data");
   }
 
   @Test
-  public void loadUri_loadsImage() throws Exception {
+  public void load_httpUri_loadsImage() throws Exception {
     SimpleBitmapLoader bitmapLoader =
         new SimpleBitmapLoader(MoreExecutors.newDirectExecutorService());
     MockWebServer mockWebServer = new MockWebServer();
@@ -93,7 +102,7 @@ public class SimpleBitmapLoaderTest {
   }
 
   @Test
-  public void loadUri_serverError_throwsException() {
+  public void load_httpUriAndServerError_throwsException() {
     SimpleBitmapLoader bitmapLoader =
         new SimpleBitmapLoader(MoreExecutors.newDirectExecutorService());
     MockWebServer mockWebServer = new MockWebServer();
@@ -106,7 +115,36 @@ public class SimpleBitmapLoaderTest {
   }
 
   @Test
-  public void loadUri_nonHttpUri_throwsException() {
+  public void load_fileUri_loadsImage() throws Exception {
+    byte[] imageData =
+        TestUtil.getByteArray(ApplicationProvider.getApplicationContext(), TEST_IMAGE_PATH);
+    File file = tempFolder.newFile();
+    Files.write(Paths.get(file.getAbsolutePath()), imageData);
+    Uri uri = Uri.fromFile(file);
+    SimpleBitmapLoader bitmapLoader =
+        new SimpleBitmapLoader(MoreExecutors.newDirectExecutorService());
+
+    Bitmap bitmap = bitmapLoader.loadBitmap(uri).get();
+
+    assertThat(
+            bitmap.sameAs(
+                BitmapFactory.decodeByteArray(imageData, /* offset= */ 0, imageData.length)))
+        .isTrue();
+  }
+
+  @Test
+  public void fileUriWithFileNotExisting() throws Exception {
+    SimpleBitmapLoader bitmapLoader =
+        new SimpleBitmapLoader(MoreExecutors.newDirectExecutorService());
+
+    assertException(
+        () -> bitmapLoader.loadBitmap(Uri.parse("file:///not_valid/path/image.bmp")).get(),
+        IllegalArgumentException.class,
+        /* messagePart= */ "Could not read image from file");
+  }
+
+  @Test
+  public void load_unhandledUriScheme_throwsException() {
     SimpleBitmapLoader bitmapLoader =
         new SimpleBitmapLoader(MoreExecutors.newDirectExecutorService());
 
@@ -114,10 +152,6 @@ public class SimpleBitmapLoaderTest {
         () -> bitmapLoader.loadBitmap(Uri.parse("/local/path")).get(),
         MalformedURLException.class,
         /* messagePart= */ "no protocol");
-    assertException(
-        () -> bitmapLoader.loadBitmap(Uri.parse("file://local/path")).get(),
-        UnsupportedOperationException.class,
-        /* messagePart= */ "Unsupported scheme");
     assertException(
         () -> bitmapLoader.loadBitmap(Uri.parse("asset://asset/path")).get(),
         MalformedURLException.class,